+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Release Notes
 All notable changes to this project will be documented in this file.
 
+## 0.20.100 - Feb 17, 2020
+### New
+- `registerAccessKey` parameters passed as structure. 
+- `restrictToAccounts` option for access keys.
+- `parseMessage` function for parsing message BOC into JSON.
+- `deploy` and `processDeployMessage` functions now check the account state before sending message and return `alreadyDeployed = true` if account is already active.
+
 ## 0.20.2 - Feb 12, 2020
 ### Fixed
 - `tracer` config parameter caused setup error (circular JSON).